pyproject.toml
[tool.poetry]
name = "reallife-quest-murder-game"
version = "1.0.0"
description = ""
authors = ["Nimuedaertya <saluneriya@googlemail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
pyyaml = "^6.0.1"
schematics = "^2.1.1"
flask-classful = "^0.16.0"
flask-assets = "^2.1.0"
argparse = "^1.4.0"
segno = "^1.6.1"
reportlab = "^4.2.0"
flask-socketio = "^5.3.6"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
flake8 = "^7.0.0"
pylint = "^3.2.2"
coverage = "^7.5.1"


[tool.coverage.run]
source = ["src/."]

[tool.pytest.ini_options]
addopts = "-p no:warnings"
pythonpath = "src"
minversion = "6.0"
testpaths = [
    "tests/",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
